ST/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework ==3.16.1
djangorestframework_simplejwt== 5.5.1
pip                        ==   25.2
PyJWT                     ==    2.10.1
sqlparse                  ==    0.5.3
tzdata                    ==    2025.2

markdown
django-filter
drf-spectacular

redis
celery
django-redis
django-celery-results
django-celery-beat
kombu


ollama